+++ contracts/ModuleRegistry.sol
@@ -4,8 +4,7 @@ import "./interfaces/IModuleRegistry.sol";
 import "./interfaces/IModuleFactory.sol";
 import "./interfaces/ISecurityToken.sol";
 import "./interfaces/ISecurityTokenRegistry.sol";
-import "./interfaces/IRegistry.sol";
-import "openzeppelin-solidity/contracts/ownership/Ownable.sol";
+import "./Registry.sol";
 
 /**
 * @title ModuleRegistry
@@ -13,7 +12,7 @@ import "openzeppelin-solidity/contracts/ownership/Ownable.sol";
 * Anyone can register modules, but only those "approved" by Polymath will be allowed to everyone.
 */
 
-contract ModuleRegistry is IModuleRegistry, IRegistry {
+contract ModuleRegistry is IModuleRegistry, Registry {
 
     // Mapping used to hold the type of module factory corresponds to the address of the Module factory contract
     mapping (address => uint8) public registry;
@@ -26,7 +25,6 @@ contract ModuleRegistry is IModuleRegistry, IRegistry {
     // Contains the list of the available tags corresponds to the module type
     mapping (uint8 => bytes32[]) public availableTags;
 
-    address public securityTokenRegistry;
     // Emit when Module been used by the securityToken
     event LogModuleUsed(address indexed _moduleFactory, address indexed _securityToken);
     // Emit when the Module Factory get registered with the ModuleRegistry contract
@@ -40,7 +38,7 @@ contract ModuleRegistry is IModuleRegistry, IRegistry {
     */
     function useModule(address _moduleFactory) external {
         //If caller is a registered security token, then register module usage
-        if (ISecurityTokenRegistry(securityTokenRegistry).isSecurityToken(msg.sender)) {
+        if (ISecurityTokenRegistry(getAddress("SecurityTokenRegistry")).isSecurityToken(msg.sender)) {
             require(registry[_moduleFactory] != 0, "ModuleFactory type should not be 0");
             //To use a module, either it must be verified, or owned by the ST owner
             require(verified[_moduleFactory]||(IModuleFactory(_moduleFactory).owner() == ISecurityToken(msg.sender).owner()),
@@ -54,7 +52,7 @@ contract ModuleRegistry is IModuleRegistry, IRegistry {
     * @dev Called by moduleFactory owner to register new modules for SecurityToken to use
     * @param _moduleFactory is the address of the module factory to be registered
     */
-    function registerModule(address _moduleFactory) external returns(bool) {
+    function registerModule(address _moduleFactory) external whenNotPaused returns(bool) {
         require(registry[_moduleFactory] == 0, "Module factory should not be pre-registered");
         IModuleFactory moduleFactory = IModuleFactory(_moduleFactory);
         require(moduleFactory.getType() != 0, "Factory type should not equal to 0");
@@ -77,15 +75,6 @@ contract ModuleRegistry is IModuleRegistry, IRegistry {
         verified[_moduleFactory] = _verified;
         emit LogModuleVerified(_moduleFactory, _verified);
         return true;
-    }
-
-    /**
-    * @dev Called by owner to set the token registry address
-    * @param _securityTokenRegistry is the address of the token registry
-    */
-    function setTokenRegistry(address _securityTokenRegistry) public onlyOwner {
-        require(_securityTokenRegistry != address(0), "Address of securityTokenregistry should not be 0x");
-        securityTokenRegistry = _securityTokenRegistry;
     }
 
     /**