+++ contracts/Pausable.sol
@@ -1,0 +1,44 @@
+pragma solidity ^0.4.23;
+
+contract Pausable {
+
+    event Pause(uint256 _timestammp);
+    event Unpause(uint256 _timestamp);
+
+    bool public paused = false;
+
+    /**
+    * @dev Modifier to make a function callable only when the contract is not paused.
+    */
+    modifier whenNotPaused() {
+        require(!paused);
+        _;
+    }
+
+    /**
+    * @dev Modifier to make a function callable only when the contract is paused.
+    */
+    modifier whenPaused() {
+        require(paused);
+        _;
+    }
+
+   /**
+    * @dev called by the owner to pause, triggers stopped state
+    */
+    function _pause() internal {
+        require(!paused);
+        paused = true;
+        emit Pause(now);
+    }
+
+    /**
+    * @dev called by the owner to unpause, returns to normal state
+    */
+    function _unpause() internal {
+        require(paused);
+        paused = false;
+        emit Unpause(now);
+    }
+
+}